Business functions
The following business functions have been added:
1. The administrator can grant/take away the administrator rights from the user
2. The customer can place an order by choosing the size of the container (number of calories) and the desired dishes in it
3. The customer can get information about how much each dish in the container will weigh and what calorie content it will be
4. The customer can view their detailed order history
5. Registration and authentication by phone number is available for couriers
6. The courier can activate/deactivate at any time to receive/not receive new orders
7. The courier can independently receive a new order
8. The courier can get detailed information about the current order and his history of delivered orders
9. The administrator can assign all available orders to couriers available at the moment (with priority according to the time of receipt of the order)
10. The customer can confirm receipt of the order
11. The courier can close the completed order only after confirmation of its receipt from the client
12. Depending on the time for which the order was completed, it is considered delivered on time or delivered late
13. The administrator can get the performance indicators of an individual courier (work speed, percentage of orders delivered on time, average number of orders delivered per day)

+++ main/src/main/java/edu/senla/dto/TypeOfContainerDTO.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @Data
 public class TypeOfContainerDTO {
 
-    private int numberOfCalories;
+    private long numberOfCalories;
 
     private String name;
